Added description box
Added a box with a short description of the page + a link to the github repo. This box can be toggled using the tab key and is hidden by default. This box will be used to add a description of the project on the project pages.
Also rewrote the 'about' section a bit.